# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERB Platform API - connecting suppliers with retailers in Saudi Arabia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/overview:
    get:
      operationId: getDashboardOverview
      tags: [dashboard]
      summary: Get dashboard overview statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard overview data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverview"
  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent platform activity feed
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/stats:
    get:
      operationId: getPlatformStats
      tags: [dashboard]
      summary: Get platform-wide supply chain statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: In-app notifications for the caller
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/unread-count:
    get:
      operationId: getUnreadNotificationCount
      tags: [notifications]
      summary: Unread notification count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                required: [count]
...
  /invoices/analyze:
    post:
      operationId: analyzeInvoice
      tags: [invoices]
      summary: Upload a previous invoice image for AI price comparison (business owner)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceAnalyzeInput"
      responses:
        "201":
          description: Invoice analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceAnalysis"
  /invoices/mine:
    get:
      operationId: getMyInvoiceAnalyses
      tags: [invoices]
      summary: List caller's past invoice analyses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Analyses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceAnalysis"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DashboardOverview:
      type: object
      properties:
        role:
          type: string
        status:
          type: string
        company_name:
          type: string
        full_name:
          type: string
        registered_at:
          type: string
        total_suppliers:
          type: integer
        total_retailers:
          type: integer
        active_connections:
          type: integer
      required:
        - role
        - status
        - company_name
        - full_name
        - registered_at
        - total_suppliers
        - total_retailers
        - active_connections
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        message:
          type: string
        timestamp:
          type: string
        role:
          type: string
      required:
        - id
        - type
        - message
        - timestamp
        - role
    PlatformStats:
      type: object
      properties:
        total_suppliers:
          type: integer
        total_retailers:
          type: integer
        pending_approvals:
          type: integer
        active_connections:
          type: integer
        countries_served:
          type: integer
        industries_covered:
          type: integer
      required:
        - total_suppliers
        - total_retailers
        - pending_approvals
        - active_connections
        - countries_served
        - industries_covered
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
        order_id:
          type: ["integer", "null"]
        read:
          type: boolean
        created_at:
          type: string
...
    InvoiceAnalyzeInput:
      type: object
      properties:
        file_base64:
          type: string
        mime_type:
          type: string
        file_name:
          type: string
      required:
        - file_base64
        - mime_type
    InvoiceAnalysis:
      type: object
      properties:
        id:
          type: integer
        business_id:
          type: string
        file_name:
          type: ["string", "null"]
        extracted_items:
          type: string
        invoice_total:
          type: ["string", "null"]
        comparison:
          type: string
        platform_total:
          type: ["string", "null"]
        potential_saving:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        created_at:
          type: string
      required:
        - id
        - business_id
        - extracted_items
        - comparison
        - created_at